requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
Flask-SQLAlchemy==2.4.4
Flask-JWT-Extended==3.24.1
Flask-Cors==3.0.9
Flask-Bcrypt==0.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
WTForms==2.3.3
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.14.0
